# Prueba_Tecnica_Inditex/rest/src/main/resources/OpenApiRestConfiguration.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Inditex Swagger
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
paths:
  /prices:
    get:
      summary: Listar precios por filtro
      operationId: listPricesByFilter
      parameters:
        - name: date
          in: query
          description: Fecha de inicio de la consulta
          required: true
          schema:
            type: string
            format: date-time
        - name: productId
          in: query
          description: Identificador del producto
          required: true
          schema:
            type: integer
            format: int64
        - name: brandId
          in: query
          description: Identificador de la cadena
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Precios recuperados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricesRestCollection"
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    PricesRest:
      type: object
      required:
        - productId
        - brandId
        - startDate
        - endDate
        - priceList
        - price
      properties:
        productId:
          type: integer
          format: int64
        brandId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        priceList:
          type: integer
          format: int64
        price:
          type: number
          format: double
    PricesRestCollection:
      type: array
      items:
        $ref: "#/components/schemas/PricesRest"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string